Refactoring for Task Service
