+++ src/tasks/tasks.service.ts
@@ -1,72 +1,57 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
-import { Task, TaskStatus } from './task.model';
-import { v4 as uuid } from 'uuid';
+import { TaskStatus } from './task-status.enum.';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
 
 @Injectable()
 export class TasksService {
-  private tasks: Task[] = [];
-
-  getAllTasks(): Task[] {
-    return this.tasks;
-  }
-
-  getTasksWithFilters(filterDto: GetTasksFilterDto): Task[] {
-    const { status, search } = filterDto;
-    let tasks = this.getAllTasks();
-    if (status) {
-      tasks = tasks.filter((task) => task.status == status);
-    }
-
-    if (search) {
-      tasks = tasks.filter((tasks) => {
-        if (
-          tasks.title.includes(search) ||
-          tasks.description.includes(search)
-        ) {
-          return true;
-        }
-
-        return false;
-      });
-    }
-
-    return tasks;
-  }
-
-  getTaskByID(id: string): Task {
-    const found = this.tasks.find((task) => task.id === id);
-
-    if (!found) {
-      throw new NotFoundException(`Task with ID "${id}" not found`);
-    }
-
-    return found;
-  }
-
-  createTask(createTaskDto: CreateTaskDto): Task {
-    const { title, description } = createTaskDto;
-
-    const task: Task = {
-      id: uuid(),
-      title,
-      description,
-      status: TaskStatus.OPEN,
-    };
-
-    this.tasks.push(task);
-    return task;
-  }
-
-  deleteTask(id: string): void {
-    const found = this.getTaskByID(id);
-    this.tasks = this.tasks.filter((tasks) => tasks.id !== found.id);
-  }
-
-  updateTaskStatus(id: string, status: TaskStatus) {
-    const task = this.getTaskByID(id);
-    task.status = status;
-    return task;
-  }
+  // getAllTasks(): Task[] {
+  //   return this.tasks;
+  // }
+  // getTasksWithFilters(filterDto: GetTasksFilterDto): Task[] {
+  //   const { status, search } = filterDto;
+  //   let tasks = this.getAllTasks();
+  //   if (status) {
+  //     tasks = tasks.filter((task) => task.status == status);
+  //   }
+  //   if (search) {
+  //     tasks = tasks.filter((tasks) => {
+  //       if (
+  //         tasks.title.includes(search) ||
+  //         tasks.description.includes(search)
+  //       ) {
+  //         return true;
+  //       }
+  //       return false;
+  //     });
+  //   }
+  //   return tasks;
+  // }
+  // getTaskByID(id: string): Task {
+  //   const found = this.tasks.find((task) => task.id === id);
+  //   if (!found) {
+  //     throw new NotFoundException(`Task with ID "${id}" not found`);
+  //   }
+  //   return found;
+  // }
+  // createTask(createTaskDto: CreateTaskDto): Task {
+  //   const { title, description } = createTaskDto;
+  //   const task: Task = {
+  //     id: uuid(),
+  //     title,
+  //     description,
+  //     status: TaskStatus.OPEN,
+  //   };
+  //   this.tasks.push(task);
+  //   return task;
+  // }
+  // deleteTask(id: string): void {
+  //   const found = this.getTaskByID(id);
+  //   this.tasks = this.tasks.filter((tasks) => tasks.id !== found.id);
+  // }
+  // updateTaskStatus(id: string, status: TaskStatus) {
+  //   const task = this.getTaskByID(id);
+  //   task.status = status;
+  //   return task;
+  // }
 }
